Personal changes
Some of the changes that I made are the basic HTML of some things. I pretty much remade the entirety of the Examples.html. I don't even know if I really changed much, there are a ton of problems that I do not know how to solve, so I will have to work on that either later today or tomorrow, whatever works best.